Rename: styles 하위 파일명, 경로 수정
